+++ .dockerignore
@@ -1,4 +1,8 @@
-/*.local
+/.git/
+/.github/
+/.env.local
+/.env.*.local
+/Dockerfile
 
 # Ignore generated credentials from google-github-actions/auth
 gha-creds-*.json
