+++ .env
@@ -1,0 +1,4 @@
+PORT=3000
+
+ETHERSCAN_API_KEY=
+HEALTHCHECK_URL=
