+++ .github/CODEOWNERS
@@ -1,2 +1,3 @@
 # CI/CD
 /.github/ @RequestFinance/ops
+Dockerfile @RequestFinance/ops
