+++ .github/workflows/ci.yaml
@@ -1,30 +1,23 @@
+name: CI
+
 on:
   push:
-    branches:
-      - main
+    branches: [main]
   pull_request:
-    branches:
-      - main
+    branches: [main]
 
 concurrency:
-  group: "${{ github.workflow }}-${{ github.event.pull_request.number || github.ref }}"
+  group: ${{ github.workflow }}-${{ github.event.pull_request.number || github.ref }}
   cancel-in-progress: true
 
 jobs:
-  build-test:
+  test:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7.0.0
-      - run: corepack enable
-      - name: Use Node.js
-        uses: actions/setup-node@48b55a011bda9f5d6aeb4c2d9c7362e8dae4041e # v6.4.0
+      - uses: denoland/setup-deno@667a34cdef165d8d2b2e98dde39547c9daac7282 # v2.0.4
         with:
-          node-version-file: ".nvmrc"
-          cache: "pnpm"
-      - run: corepack enable
-      - name: Install deps
-        run: pnpm i
-      - name: Build
-        run: pnpm build
-      - name: Test
-        run: pnpm test
+          deno-version: 2.8.2
+      - run: deno fmt --check
+      - run: deno lint
+      - run: deno task test
