+++ .hooks/pre-commit
@@ -1,0 +1,10 @@
+#!/usr/bin/env sh
+. "$(dirname -- "$0")/_/hook.sh"
+
+branch="$(git rev-parse --abbrev-ref HEAD)"
+if [ "$branch" = "master" ] || [ "$branch" = "main" ]; then
+  echo "You can't commit directly to $branch branch"
+  exit 1
+fi
+
+deno run -A npm:lint-staged@15.5.0
