--- .nvmrc
@@ -1,1 +1,0 @@
-v20
