+++ deno.json
@@ -1,0 +1,19 @@
+{
+  "imports": {
+    "@oak/oak": "jsr:@oak/oak@^17.1.6",
+    "@oak/commons": "jsr:@oak/commons@^1.0.0",
+    "@std/expect": "jsr:@std/expect@^1.0.13",
+    "@std/testing": "jsr:@std/testing@^1.0.9",
+    "~/": "./src/",
+    "dotenv-flow": "npm:dotenv-flow@^4.1.0",
+    "env-var": "npm:env-var@^7.5.0",
+    "gsheet-object": "npm:gsheet-object@0.1.6"
+  },
+  "tasks": {
+    "start": "deno run --allow-read --allow-env --allow-net --allow-sys src/index.ts",
+    "dev": "deno run --watch --allow-read --allow-env --allow-net --allow-sys src/index.ts",
+    "hook": "deno run --allow-read --allow-run --allow-write https://deno.land/x/deno_hooks@0.1.1/mod.ts",
+    "test": "NODE_ENV=test deno test --allow-read --allow-env --allow-net --allow-sys tests"
+  },
+  "minimumDependencyAge": "P7D"
+}
