--- package.json
@@ -1,33 +1,0 @@
-{
-  "name": "wallet-balance-notifier",
-  "private": "true",
-  "dependencies": {
-    "axios": "^1.13.5",
-    "dotenv": "^16.4.5",
-    "googleapis": "^134.0.0",
-    "gsheet-object": "0.1.6",
-    "web3-utils": "^4.2.1"
-  },
-  "devDependencies": {
-    "@types/jest": "^29.5.12",
-    "@types/node": "^20.11.30",
-    "jest": "^29.7.0",
-    "prettier": "^3.2.5",
-    "rimraf": "^5.0.5",
-    "ts-jest": "^29.1.2",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.4.3"
-  },
-  "scripts": {
-    "build": "rimraf build && mkdir build && cp package.json build/ && tsc",
-    "watch": "tsc --watch",
-    "test": "jest",
-    "prettier": "prettier --write .",
-    "start": "ts-node src/run.ts"
-  },
-  "engines": {
-    "node": "^20.0.0",
-    "pnpm": ">=8.0.0"
-  },
-  "packageManager": "pnpm@8.14.3"
-}
