--- src/run.ts
@@ -1,5 +1,0 @@
-import { config } from "dotenv";
-import { main } from ".";
-
-config();
-void main();
