--- tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "compilerOptions": {
-    "lib": ["ES2023"],
-    "module": "commonjs",
-    "target": "ES2022",
-    "strict": true,
-    "esModuleInterop": true,
-    "rootDir": "src",
-    "outDir": "build",
-    "baseUrl": "./",
-    "paths": {
-      "*": ["src/typings/*"]
-    }
-  },
-  "include": ["src/**/*"]
-}
